reorganizing data loading
