+++ src/tasks/tasks.py
@@ -29,8 +29,9 @@ from allennlp.data.fields import TextField, LabelField, \
     SpanField, ListField, MetadataField
 from ..allennlp_mods.numeric_field import NumericField
 
-from ..utils import utils, tokenizers
-from ..utils.utils import load_tsv, process_sentence, truncate, load_diagnostic_tsv
+from ..utils import utils
+from ..utils.utils import truncate
+from ..utils.data_loaders import load_tsv, process_sentence, load_diagnostic_tsv
 
 from typing import Iterable, Sequence, List, Dict, Any, Type
 
@@ -139,7 +140,7 @@ class Task(object):
 
     def tokenizer_is_supported(self, tokenizer_name):
         ''' Check if the tokenizer is supported for this task. '''
-        return tokenizer_name in tokenizers.AVAILABLE_TOKENIZERS
+        return tokenizer_name in tokenizers.AVAILABLE_TOKENIZERS.keys()
 
     @property
     def tokenizer_name(self):
